backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
hatchet-sdk
hatchet-server
pdfplumber
requests
reportlab
